add Cucumber featcher test
