+++ IvanVaskov/homework02/converter/features/support/env.rb
@@ -1,2 +1,3 @@
 require 'cucumber/rspec/doubles'
 require 'converter'
+require 'money'
