+++ IvanVaskov/homework02/main/data.csv
@@ -1,27 +1,27 @@
 Cur_Abbreviation,Cur_Scale,Cur_Name,Cur_OfficialRate
-AUD,1,Австралийский доллар,1.7793
-BGN,1,Болгарский лев,1.504
-UAH,100,Гривен,9.0526
-DKK,10,Датских крон,3.9505
-USD,1,Доллар США,2.4856
-EUR,1,Евро,2.9459
-PLN,10,Злотых,6.695
-IRR,100000,Иранских риалов,5.9181
-ISK,100,Исландских крон,1.8169
-JPY,100,Йен,2.3461
-CAD,1,Канадский доллар,1.8813
-CNY,10,Китайских юаней,3.5914
-KWD,1,Кувейтский динар,8.1261
-MDL,10,Молдавских леев,1.5014
-NZD,1,Новозеландский доллар,1.624
-NOK,10,Норвежских крон,2.7796
-RUB,100,Российских рублей,3.3736
-XDR,1,СДР (Специальные права заимствования),3.5252
-SGD,1,Сингапурcкий доллар,1.8142
-KGS,100,Сомов,3.1908
-KZT,1000,Тенге,5.9233
-TRY,10,Турецких лир,3.3926
-GBP,1,Фунт стерлингов,3.2604
-CZK,100,Чешских крон,11.2823
-SEK,10,Шведских крон,2.8466
-CHF,1,Швейцарский франк,2.7254
+AUD,1,Австралийский доллар,1.8661
+BGN,1,Болгарский лев,1.5741
+UAH,100,Гривен,9.4958
+DKK,10,Датских крон,4.1368
+USD,1,Доллар США,2.6028
+EUR,1,Евро,3.0758
+PLN,10,Злотых,7.0048
+IRR,100000,Иранских риалов,6.1971
+ISK,100,Исландских крон,1.8912
+JPY,100,Йен,2.4473
+CAD,1,Канадский доллар,1.9695
+CNY,10,Китайских юаней,3.7655
+KWD,1,Кувейтский динар,8.5092
+MDL,10,Молдавских леев,1.5632
+NZD,1,Новозеландский доллар,1.7004
+NOK,10,Норвежских крон,2.9067
+RUB,100,Российских рублей,3.4939
+XDR,1,СДР (Специальные права заимствования),3.6769
+SGD,1,Сингапурcкий доллар,1.9011
+KGS,100,Сомов,3.3412
+KZT,1000,Тенге,6.2085
+TRY,10,Турецких лир,3.5242
+GBP,1,Фунт стерлингов,3.4134
+CZK,100,Чешских крон,11.7945
+SEK,10,Шведских крон,2.9728
+CHF,1,Швейцарский франк,2.8627
